Added better protection to routes.

+++ src/controllers/comment.controller.ts
@@ -127,15 +127,27 @@ export const updateComment = [
 			return res.status(400).json({ errors: errors.array() });
 		}
 
-		try {
-			const { content } = req.body;
-			const comment = await Comment.findByIdAndUpdate(
-				req.params.id,
-				{ content },
-				{ new: true },
-			);
+		const user = req.user as IUser;
+		if (!user) {
+			return res.status(401).json({ message: "Unauthorized" });
+		}
 
-			res.status(201).json(comment);
+		try {
+			const comment = await Comment.findById(req.params.id);
+
+			if (!comment) {
+				return res.status(404).json({ message: "Comment not found" });
+			}
+
+			if (comment.author.toString() !== user._id.toString()) {
+				return res.status(403).json({ message: "Not authorized" });
+			}
+			const { content } = req.body;
+
+			comment.content = content;
+			await comment.save();
+
+			res.status(200).json(comment);
 		} catch (error) {
 			res.status(400).json({ message: error.message });
 		}
@@ -148,14 +160,26 @@ export const updateComment = [
 export const deleteComment = [
 	passport.authenticate("jwt", { session: false }),
 	expressAsyncHandler(async (req: Request, res: Response): Promise<any> => {
+		const user = req.user as IUser;
+
+		if (!user) {
+			return res.status(401).json({ message: "Unauthorized" });
+		}
+
 		try {
-			const comment = await Comment.findByIdAndDelete(req.params.id);
+			const comment = await Comment.findById(req.params.id);
 
 			if (!comment) {
 				return res.status(404).json({ message: "Comment not found" });
 			}
 
-			res.status(201).json(comment);
+			if (comment.author.toString() !== user._id.toString()) {
+				return res.status(403).json({ message: "Not authorized" });
+			}
+
+			await Comment.findByIdAndDelete(req.params.id);
+
+			res.status(200).json({ message: "Comment deleted successfully" });
 		} catch (error) {
 			res.status(500).json({ message: error.message });
 		}
@@ -164,7 +188,7 @@ export const deleteComment = [
 
 // @desc    Get comments by post id
 // @route   GET posts/:id/comments
-// @access  Private
+// @access  Public
 export const getCommentsByPostId = expressAsyncHandler(
 	async (req: Request, res: Response) => {
 		try {
